start breaking ui components out
