+++ src/app/products/ecr-viewer/page.tsx
@@ -7,13 +7,19 @@ import {
   ProcessListHeading,
   Accordion,
   Link,
-  SideNav,
 } from '@trussworks/react-uswds';
-import classNames from 'classnames';
 import Image from 'next/image';
-import { SetStateAction, useState } from 'react';
 import './styles.scss';
 import { RoundedBackground } from '@/app/components/RoundedBackground/RoundedBackground';
+import {
+  Navigation,
+  TitleSection,
+  SectionHeader,
+  SectionSubheader,
+  AccordionItemContent,
+  SendMailLink,
+  GithubNav,
+} from '../_ui';
 
 export default function EcrViewer() {
   return (
@@ -129,7 +135,7 @@ export default function EcrViewer() {
                     your jurisdiction.
                   </p>
                 </div>
-                <div className="flex flex-col gap-8">
+                <div className="flex flex-col gap-3">
                   <SectionSubheader>eCR Summary</SectionSubheader>
                   <p className="m-0 p-0">
                     The eCR Viewer surfaces key information in a summary at the
@@ -145,7 +151,7 @@ export default function EcrViewer() {
                   height={383}
                   src={`${basePath}/images/products/ecr-viewer/how-it-works.gif`}
                 />
-                <div className="flex flex-col gap-6">
+                <div className="flex flex-col gap-3">
                   <SectionSubheader>Combined RR and eICR</SectionSubheader>
                   <p className="m-0 flex flex-col gap-8 p-0">
                     <span>
@@ -164,7 +170,7 @@ export default function EcrViewer() {
                     </span>
                   </p>
                 </div>
-                <div className="flex flex-col gap-6">
+                <div className="flex flex-col gap-3">
                   <SectionSubheader>Consistent data display</SectionSubheader>
                   <p className="m-0 p-0">
                     No matter who sent the eCR, every eCR you view in the eCR
@@ -563,164 +569,5 @@ export default function EcrViewer() {
         </div>
       </ContentContainer>
     </div>
-  );
-}
-
-interface AccordionItemContentProps {
-  children: React.ReactNode;
-}
-
-const AccordionItemContent = ({ children }: AccordionItemContentProps) => {
-  return (
-    <span className="font-['Public Sans'] text-base font-normal leading-snug text-[#224a58]">
-      {children}
-    </span>
-  );
-};
-
-function Navigation() {
-  const [selectedHash, setSelectedHash] = useState('#overview');
-
-  return (
-    <SideNav
-      items={[
-        <NavItem
-          title="Overview"
-          id="overview"
-          key="overview"
-          selectedHash={selectedHash}
-          setSelectedHash={setSelectedHash}
-        />,
-        <NavItem
-          title="Product features"
-          id="product-features"
-          key="product-features"
-          selectedHash={selectedHash}
-          setSelectedHash={setSelectedHash}
-        />,
-        <NavItem
-          title="How it works"
-          id="how-it-works"
-          key="how-it-works"
-          selectedHash={selectedHash}
-          setSelectedHash={setSelectedHash}
-        />,
-        <NavItem
-          title="Getting started"
-          id="getting-started"
-          key="getting-started"
-          selectedHash={selectedHash}
-          setSelectedHash={setSelectedHash}
-        />,
-        <NavItem
-          title="Technical resources"
-          id="technical-resources"
-          key="technical-resources"
-          selectedHash={selectedHash}
-          setSelectedHash={setSelectedHash}
-        />,
-        <NavItem
-          title="FAQs"
-          id="faqs"
-          key="faqs"
-          selectedHash={selectedHash}
-          setSelectedHash={setSelectedHash}
-        />,
-      ]}
-    />
-  );
-}
-
-interface NavItemProps {
-  title: string;
-  id: string;
-  selectedHash: string;
-  setSelectedHash: (value: SetStateAction<string>) => void;
-}
-function NavItem({ title, id, selectedHash, setSelectedHash }: NavItemProps) {
-  const itemHash = `#${id}`;
-  return (
-    <a
-      href={itemHash}
-      key={id}
-      className={classNames({
-        'usa-current': itemHash === selectedHash,
-      })}
-      onClick={() => setSelectedHash(itemHash)}
-    >
-      {title}
-    </a>
-  );
-}
-
-interface GithubNavProps {
-  version: string;
-  githubHref: string;
-}
-
-function GithubNav({ version, githubHref }: GithubNavProps) {
-  return (
-    <div className="max-w-[16.25rem] bg-[#e7f2f5] pb-5 pt-7">
-      <div className="mr-10 flex flex-col items-center gap-1">
-        <p className="m-0 p-0 text-base font-normal leading-relaxed text-[#224a58]">
-          Current version: {version}
-        </p>
-        <hr className="ml-10 h-[0px] w-full border border-[#a9aeb1]" />
-        <Link
-          href={githubHref}
-          className="text-base font-semibold leading-relaxed text-[#3a7d95] underline"
-        >
-          GitHub Respository
-        </Link>
-      </div>
-    </div>
-  );
-}
-
-interface SectionHeader {
-  children: React.ReactNode;
-}
-
-const SectionHeader = ({ children }: SectionHeader) => {
-  return (
-    <h2 className="font-['Public Sans'] text-[1.75rem] font-bold leading-[2.4rem] text-[#14333d]">
-      {children}
-    </h2>
-  );
-};
-
-const SectionSubheader = ({ children }: SectionHeader) => {
-  return (
-    <h3 className="font-['Public Sans'] text-xl font-bold leading-7 text-[#224a58]">
-      {children}
-    </h3>
-  );
-};
-
-interface TitleSection {
-  title: string;
-  text: string;
-}
-function TitleSection({ title, text }: TitleSection) {
-  return (
-    <div className="flex flex-col gap-2">
-      <h1 className="font-['Merriweather'] text-[2rem] font-bold leading-[2.8rem] text-[#224a58]">
-        {title}
-      </h1>
-      <p className="font-['Public Sans'] m-0 p-0 text-[1.38rem] font-extralight leading-[2rem] text-[#224a58]">
-        {text}
-      </p>
-    </div>
-  );
-}
-
-function SendMailLink() {
-  return (
-    <Link
-      className="font-['Public Sans'] text-base font-bold leading-snug text-[#3a7d95]"
-      href="mailto:dibbs@cdc.gov"
-    >
-      dibbs@cdc.gov
-    </Link>
   );
 }
